orchestrator/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
protobuf==4.25.2
Werkzeug==3.0.1
Flask-CORS==4.0.0
jsonschema==4.21.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0